requirements.txt
Django>=4.0.0
django-crispy-forms
Pillow
django-allauth
gunicorn
whitenoise
psycopg2-binary
python-dotenv
django-jazzmin
msal
requests 